dDevice-api/requirements.txt
mysqlclient
django
djangorestframework
markdown
django-filter
django-cors-headers
django-filter
